Convert training data to inputs for Python script #12
The Python script can't deal yet with different numbers of inputs, that has to come later.
The new CSV files all have commas as delimiter, no semicolons anymore.
Also the inputs CSV files don't have headers anymore, they are not meant to be kept for longer, in contrast to the CSV files with position IDs.