Increase performance of mixed move generation by 24 percent
When finding checkers to move we have to iterate over pips. We now don't iterate over pips which would result in a bearoff when that bearoff is currently illegal. Because of that we also can also use a faster `can_move` method which does not check whether bearoffs are legal.