Remove struct `CompositeEvaluator`
`CompositeEvaluator` was more elegant and general than `ComplexEvaluator`, but only able to evaluate a single position at once.
The neural net inference is roughly three times as fast when it can work on multiple positions at once; so single position evaluation is just not practical.

The `GameOverEvaluator` was moved to `ComplexEvaluator` as it's not needed anywhere else.